refactor(2161): Rename user action step widgets
